+++ kinto/core/storage/postgresql/__init__.py
@@ -65,7 +65,7 @@ class Storage(StorageBase):
 
     """  # NOQA
 
-    schema_version = 14
+    schema_version = 15
 
     def __init__(self, client, max_fetch_size, *args, **kwargs):
         super().__init__(*args, **kwargs)