+++ tests/core/test_views_transaction.py
@@ -1,5 +1,8 @@
 import mock
+import time
+import threading
 import unittest
+from uuid import uuid4
 
 from pyramid import testing
 from pyramid import httpexceptions
@@ -120,6 +123,44 @@ class TransactionTest(PostgreSQLTest, unittest.TestCase):
 
         resp = self.app.get('/psilos', headers=self.headers)
         self.assertEqual(len(resp.json['data']), 0)
+
+
+class IntegrityConstraintTest(PostgreSQLTest, unittest.TestCase):
+    def get_app_settings(self, extras=None):
+        settings = super().get_app_settings(extras)
+        if sqlalchemy is not None:
+            settings.pop('storage_poolclass', None)  # Use real pool.
+        return settings
+
+    def test_concurrent_transactions_do_not_fail(self):
+        # This test originally intended to reproduce integrity errors and check
+        # that a 409 was obtained. But since every errors that could be reproduced
+        # could be also be fixed, this test just asserts that API responses
+        # are consistent. # See Kinto/kinto#1125.
+
+        # Make requests slow.
+        patch = mock.patch('kinto.core.resource.UserResource.postprocess',
+                           lambda s, r, a='read', old=None: time.sleep(0.2) or {})
+        patch.start()
+        self.addCleanup(patch.stop)
+
+        # Same object created in two concurrent requests.
+        body = {'data': {'id': str(uuid4())}}
+        results = set()
+
+        def create_object():
+            r = self.app.post_json('/psilos', body, headers=self.headers,
+                                   status=(201, 200, 409))
+            results.add(r.status_code)
+
+        thread1 = threading.Thread(target=create_object)
+        thread2 = threading.Thread(target=create_object)
+        thread1.start()
+        time.sleep(0.1)
+        thread2.start()
+        thread1.join()
+        thread2.join()
+        self.assertTrue({201, 200, 409} >= results)
 
 
 @skip_if_no_postgresql